pyproject.toml
[tool.poetry]
name = "llm-ocr"
version = "0.1.0"
description = ""
authors = ["hrtaym1114-github <98378300+hrtaym1114-github@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyautogui = "^0.9.54"
python-dotenv = "^1.0.1"
openai = "^1.53.0"
pillow = "^11.0.0"
pyperclip = "^1.9.0"
pyqt6 = "^6.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
